Converted gradle to use the 'application' plugin. Configured jar gradle task so we can build jars.
