+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'siren'
+rootProject.name = 'soscmd'
