+++ src/main/java/com/sirenofshame/soscmd/SirenConstants.java
@@ -1,8 +1,3 @@
-/*
- * Copyright (c) 2015.
- * Filip Hanik
- */
-
 package com.sirenofshame.soscmd;
 
 public class SirenConstants {
