+++ src/main/resources/javax.usb.properties
@@ -1,6 +1,1 @@
-#
-# Copyright (c) 2015.
-# Filip Hanik
-#
-
 javax.usb.services = org.usb4java.javax.Services
